Added interface and classes for db connection
