+++ src/main/java/pl/kowalczyk/maciej/java/app/bookingapp/dao/repository/AddressRepository.java
@@ -1,0 +1,7 @@
+package pl.kowalczyk.maciej.java.app.bookingapp.dao.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import pl.kowalczyk.maciej.java.app.bookingapp.dao.repository.entity.AddressEntity;
+
+public interface AddressRepository extends JpaRepository<AddressEntity, Long> {
+}
